+++ src/ui/CanvasHelpers.h
@@ -11,14 +11,14 @@ inline void fill(ftxui::Canvas& c, int x, int y, int w, int h, ftxui::Color bg) 
     if (row * 4 >= c.height()) break;
     int len = std::min(w, (c.width() - x * 2) / 2);
     if (len <= 0) break;
-    c.DrawText(x * 2, row * 4, std::string(len, ' '),
+    c.DrawText(x * 2, row * 4, std::string(static_cast<size_t>(len), ' '),
       [bg](ftxui::Pixel& p) { p.background_color = bg; });
   }
 }
 
 inline void write(ftxui::Canvas& c, int x, int y, const std::string& text, ftxui::Color fg, ftxui::Color bg) {
   int maxLen = (c.width() - x * 2) / 2;
-  auto t = (int)text.size() > maxLen ? text.substr(0, maxLen) : text;
+  auto t = static_cast<int>(text.size()) > maxLen ? text.substr(0, static_cast<size_t>(maxLen)) : text;
   c.DrawText(x * 2, y * 4, t, [fg, bg](ftxui::Pixel& p) {
     p.foreground_color = fg;
     p.background_color = bg;
